services/authlinc-api/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.6.1
email-validator==2.1.0
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.1
structlog==24.1.0

